refactor (events): add events screen and integrate navigation in drawer
