+++ app/(tabs)/events.tsx
@@ -1,0 +1,2 @@
+export { EventsScreen as default } from "@/src/features/events/screens/EventsScreen";
+
